+++ MeetingBar/Extensions/String.swift
@@ -69,11 +69,19 @@ extension String {
             html: data,
             options: [.documentType: NSAttributedString.DocumentType.html],
             documentAttributes: nil
-        ) {
+           ) {
             return attributedSelf.string
         }
         return self
     }
+
+    func fileName() -> String {
+        URL(fileURLWithPath: self).deletingPathExtension().lastPathComponent
+    }
+
+    func fileExtension() -> String {
+        URL(fileURLWithPath: self).pathExtension
+	}
 
     func encodeUrl() -> String? {
         self.addingPercentEncoding( withAllowedCharacters: NSCharacterSet.urlQueryAllowed)
@@ -140,5 +148,27 @@ extension String {
         let output = self.splitWithNewLineString(with: attributes, maxWidth: maxWidth)
         let attributedString = NSAttributedString(string: output, attributes: attributes)
         return attributedString
+    }
+}
+
+extension NSAttributedString {
+    func withLinksEnabled() -> NSAttributedString {
+        let linkDetectionRegexPattern = #"(http|ftp|https)://([\w_-]+(?:(?:\.[\w_-]+)+))([\w.,@?^=%&:/~+#-]*[\w@?^=%&/~+#-])?"#
+        guard let regex = try? NSRegularExpression(pattern: linkDetectionRegexPattern, options: .caseInsensitive) else {
+            return self
+        }
+
+        let newAttributedString = NSMutableAttributedString(attributedString: self)
+        for match in regex.matches(in: self.string, range: NSRange(location: 0, length: self.string.utf16.count)) {
+            guard let range = Range(match.range, in: self.string) else {
+                continue
+            }
+            let urlString = String(self.string[range])
+            guard let url = URL(string: urlString) else {
+                continue
+            }
+            newAttributedString.addAttribute(.link, value: url, range: match.range)
+        }
+        return NSAttributedString(attributedString: newAttributedString)
     }
 }