+++ MeetingBar/Views/Changelog/Changelog.swift
@@ -26,6 +26,42 @@ struct ChangelogView: View {
                         Text("and small bug fixes")
                     }
                 }
+                if lastRevisedVersionInChangelog < "3.3.0" {
+                    Section(header: Text("Version 3.3.0")) {
+                        Text("⏱️ Fixed bug with timer freeze")
+                        Text("🧰 Browser management")
+                        Text("⚡ Quick Actions: ")
+                        Text("  - Show/hide meeting title in status bar")
+                        Text("  - Open meeting from clipboard")
+                        Text("• Customizable appereance for events without meeting links")
+                        Text("• Localization")
+                        Text("• Create meetings in Jam")
+                        Text("• Open event in Fantastical from event submenu")
+                        Text("• Integration with subscribed calendars")
+                    }
+                }
+                if lastRevisedVersionInChangelog < "3.4.0" {
+                    Section(header: Text("Version 3.4.0")) {
+                        Text("📋 New view of notes in the event submenu with selectable text and clickable links.")
+                        Text("🧭 Fixed a bug with opening meetings in a new browser instance")
+                        Text("and small bug fixes")
+                    }
+                }
+                if lastRevisedVersionInChangelog < "3.5.0" {
+                    Section(header: Text("Version 3.5.0")) {
+                        Text("🌍 Added translations into Croatian, German, French, and Norwegian Bokmål")
+                        Text("All app notifications are now removed after all meetings are over")
+                        Text("Improved RingCentral and Zoom links detection")
+                        Text("and small bug fixes")
+                    }
+                }
+                if lastRevisedVersionInChangelog < "3.6.0" {
+                    Section(header: Text("Version 3.6.0")) {
+                        Text("🌍 Added translations into Czech")
+                        Text("Added integration with Vowel")
+                        Text("Fixed zoom link detection")
+                    }
+                }
             }.listStyle(SidebarListStyle())
             Button("Close", action: close)
         }.padding()